Add pull_request trigger
The idea here is that the pipeline will trigger when a PR is
opened/syncronised/reopened.
Should the pipeline fail, it will block the PR from being mergeable. It
would have been better to use the `issue_comment` trigger and use
comment something like `/deploy`, however this will use the primary
branch, rather than the PR branch

It is possible to run the checkout action step against the PR branch.
But the pipeline that will be run, will be that in the primary branch.
I.e. there would be no way to test a pipeline change on a feature branch
with the `issue_comment` trigger

https://github.com/orgs/community/discussions/59389