Run teardown script passing in required env vars
Note, the `URL_TO_DELETE` is currently being implied from a
parameterised string. As a more fool-proof solution, this *should* come
from the output of the `publish` pipeline to ensure the URL is exact.
Currently, we are making an assumption in the URL format that Cloudflare
are setting for preview environments